Fix ColorThief's namespace. Fix some of the issues found by CodeFactor.
