+++ Utility/ColorThief/VBox.cs
@@ -1,7 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 
-namespace BhModule.Community.Pathing.Utility {
+namespace BhModule.Community.Pathing.Utility.ColorThief {
 
     /// <summary>
     ///     3D color space box.
@@ -108,32 +107,6 @@ namespace BhModule.Community.Pathing.Utility {
             }
 
             return avg;
-        }
-
-    }
-
-    internal class VBoxCountComparer : IComparer<VBox> {
-
-        public int Compare(VBox x, VBox y) {
-            var a = x.Count(false);
-            var b = y.Count(false);
-            return a < b ? -1 : (a > b ? 1 : 0);
-        }
-
-    }
-
-    internal class VBoxComparer : IComparer<VBox> {
-
-        public int Compare(VBox x, VBox y) {
-            var aCount  = x.Count(false);
-            var bCount  = y.Count(false);
-            var aVolume = x.Volume(false);
-            var bVolume = y.Volume(false);
-
-            // Otherwise sort by products
-            var a = aCount * aVolume;
-            var b = bCount * bVolume;
-            return a < b ? -1 : (a > b ? 1 : 0);
         }
 
     }
